Add base implementation of ArduinoUSB with JSON (#9)
* Add base implementation of ArduinoUSB with JSON

This adds the base code for maintaining a JSON object state within the
Arduino. This can be used for both settings to be used (e.g. sensor
thresholds and subscription info), and status of sensors (e.g. distance,
limit state, etc.)

* Move serial state handling into class

This moves the serial state handling logic into a class for better
organization. Next will be to move it to separate files.

* Move serialstate to its own files

This moves the serialstate class to its own files to clean up the
implementation