Add base MeltdownLog and topics
This adds the MeltdownLog entry point, and a system of topics that can
be logged.

Utils: Add division-by-zero support to rounding

Refactor MeltdownLog to LogTopicRegistry

MeltdownLog was really just functioning as a topic registry, so this
moves it to the topics package so it can used more directly by topics.

LogTopicRegistry: Add static instance.

LogTopics: Update values and test

This commit fixes some faulty logic with writing out the previous edge
values.

MeltdownLog: Add LogSession

This adds the LogSession and fits it to the registry.

Move subscriptions to within topics.

This seeks to move the subscription lists to within the LogTopic base
class. It simplifies the code and also creates better performance
through fewer function chains.