PrintStreamLogWriter updates
This adds prefix and decimalPlaces to the PrintStreamLogWriter for
better formatting of the output.

prefix is displayed at the beginning of each log line and can be "" if
not desired. This is to disambiguate between multiple writers on the
same prinstream (e.g. stdout)

decimalPlaces formats floating point values that are written to truncate
past a certain decimal place, avoiding printing things like 25.20000001